6-size.c

+++ 0x00-hello_world/6-size.c
@@ -8,11 +8,11 @@
 int main(void)
 {
 
-printf("size of a char: %i byte(s)\n", (unsigned long)sizeof(char));
-printf("size of an int: %i byte(s)\n", (unsigned long)sizeof(int));
-printf("size of long int: %i bytes(s)\n", (unsigned long)sizeof(long int);
-printf("size of a long long int: %i bytes(s)\n", (unsigned long)sizeof(long long int));
-printf("size of a float : %i bytes(s)\n", (unsigned long)sizeof(float));
+printf("size of a char: %i byte(s)\n", sizeof(char));
+printf("size of an int: %i byte(s)\n", sizeof(int));
+printf("size of long int: %i bytes(s)\n", sizeof(long int);
+printf("size of a long long int: %i bytes(s)\n", sizeof(long long int));
+printf("size of a float : %i bytes(s)\n", sizeof(float));
 
 return (0);
 }
